Some docs added
