+++ src/main/kotlin/org/penakelex/database/extenstions/LocalDate.kt
@@ -2,6 +2,11 @@ package org.penakelex.database.extenstions
 
 import java.time.LocalDate
 
+/**
+ * Gets age of person on current day, using this LocalDate as date of birth
+ * @receiver [LocalDate] date of birth of the person
+ * @return [Int] - age of the person
+ * */
 fun LocalDate.getAgeOfPersonToday(): Int {
     val todayDate = LocalDate.now()
     return if (todayDate.monthValue - this.monthValue > 0
