+++ src/main/kotlin/org/penakelex/database/services/TableService.kt
@@ -7,6 +7,10 @@ import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransacti
  * Abstract table service with database query method
  * */
 abstract class TableService {
+    /**
+     * Database query executor
+     * @param query block to execute with database
+     * */
     suspend fun <Type : Any> databaseQuery(query: suspend () -> Type) =
         newSuspendedTransaction(Dispatchers.IO) { query() }
 }
