+++ src/main/kotlin/org/penakelex/routes/file/FilesControllerImplementation.kt
@@ -3,12 +3,12 @@ package org.penakelex.routes.file
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.response.*
-import org.penakelex.fileSystem.FileManager
+import org.penakelex.fileSystem.FileManagerImplementation
 import org.penakelex.response.Result
 import org.penakelex.response.toResultResponse
 
 class FilesControllerImplementation(
-    private val fileManager: FileManager
+    private val fileManager: FileManagerImplementation
 ) : FilesController {
     override suspend fun getFile(call: ApplicationCall) {
         val file = fileManager.downloadFile(
